Move wmicQuery into lib/wmic.query
